pyk/{cli/utils,utils}: promote a bunch of utilities out of cli.utils to utils
